Refactor code creating new class LargeMethodSplitter
