+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/desugar/Desugar.java
@@ -324,8 +324,6 @@ import static org.wso2.ballerinalang.compiler.desugar.ASTBuilderUtil.createState
 import static org.wso2.ballerinalang.compiler.desugar.ASTBuilderUtil.createVariable;
 import static org.wso2.ballerinalang.compiler.desugar.ASTBuilderUtil.createVariableRef;
 import static org.wso2.ballerinalang.compiler.util.CompilerUtils.getMajorVersion;
-import static org.wso2.ballerinalang.compiler.util.Constants.INIT_METHOD_SPLIT_SIZE;
-import static org.wso2.ballerinalang.compiler.util.Constants.LISTENER_COUNT_PER_METHOD;
 import static org.wso2.ballerinalang.compiler.util.Names.GENERATED_INIT_SUFFIX;
 import static org.wso2.ballerinalang.compiler.util.Names.GEN_VAR_PREFIX;
 import static org.wso2.ballerinalang.compiler.util.Names.IGNORE;
@@ -379,6 +377,7 @@ public class Desugar extends BLangNodeVisitor {
     private Unifier unifier;
     private MockDesugar mockDesugar;
     private ClassClosureDesugar classClosureDesugar;
+    private LargeMethodSplitter largeMethodSplitter;
 
     public Stack<BLangLockStmt> enclLocks = new Stack<>();
     private BLangOnFailClause onFailClause;
@@ -395,9 +394,6 @@ public class Desugar extends BLangNodeVisitor {
     private int recordCount = 0;
     private int errorCount = 0;
     private int annonVarCount = 0;
-    private int initFuncIndex = 0;
-    private int startFuncIndex = 0;
-    private int stopFuncIndex = 0;
     private int indexExprCount = 0;
     private int letCount = 0;
     private int varargCount = 0;
@@ -437,6 +433,7 @@ public class Desugar extends BLangNodeVisitor {
         this.observabilityDesugar = ObservabilityDesugar.getInstance(context);
         this.code2CloudDesugar = Code2CloudDesugar.getInstance(context);
         this.annotationDesugar = AnnotationDesugar.getInstance(context);
+        this.largeMethodSplitter = LargeMethodSplitter.getInstance(context);
         this.types = Types.getInstance(context);
         this.names = Names.getInstance(context);
         this.names = Names.getInstance(context);
@@ -691,7 +688,7 @@ public class Desugar extends BLangNodeVisitor {
      * @param bLangFunction function node
      * @param env           Symbol environment
      */
-    private void createInvokableSymbol(BLangFunction bLangFunction, SymbolEnv env) {
+    protected void createInvokableSymbol(BLangFunction bLangFunction, SymbolEnv env) {
         BType returnType = bLangFunction.returnTypeNode.getBType() == null ?
                 symResolver.resolveTypeNode(bLangFunction.returnTypeNode, env) :
                 bLangFunction.returnTypeNode.getBType();
@@ -793,9 +790,7 @@ public class Desugar extends BLangNodeVisitor {
         addNilReturnStatement((BLangBlockFunctionBody) pkgNode.stopFunction.body);
 
         if (isJvmTarget) {
-            pkgNode.initFunction = splitInitFunction(pkgNode, env);
-            pkgNode.startFunction = splitStartFunction(pkgNode, env);
-            pkgNode.stopFunction = splitStopFunction(pkgNode, env);
+            largeMethodSplitter.splitLargeFunctions(pkgNode, env);
         }
 
         pkgNode.initFunction = rewrite(pkgNode.initFunction, env);
@@ -810,7 +805,6 @@ public class Desugar extends BLangNodeVisitor {
             rewrite(testablePkg, this.symTable.pkgEnvMap.get(testablePkg.symbol));
         }
         pkgNode.completedPhases.add(CompilerPhase.DESUGAR);
-        initFuncIndex = 0;
         result = pkgNode;
     }
 
@@ -8222,7 +8216,7 @@ public class Desugar extends BLangNodeVisitor {
         return blockNode;
     }
 
-    private BLangInvocation createInvocationNode(String functionName, List<BLangExpression> args, BType retType) {
+    protected BLangInvocation createInvocationNode(String functionName, List<BLangExpression> args, BType retType) {
         BLangInvocation invocationNode = (BLangInvocation) TreeBuilder.createInvocationNode();
         BLangIdentifier name = (BLangIdentifier) TreeBuilder.createIdentifierNode();
         name.setLiteral(false);
@@ -10019,224 +10013,6 @@ public class Desugar extends BLangNodeVisitor {
         result = rewriteExpr(stmtExpr);
     }
 
-    /**
-     * Split packahe init function into several smaller functions.
-     *
-     * @param packageNode package node
-     * @param env symbol environment
-     * @return initial init function but trimmed in size
-     */
-    private BLangFunction splitInitFunction(BLangPackage packageNode, SymbolEnv env) {
-        int methodSize = INIT_METHOD_SPLIT_SIZE;
-        BLangBlockFunctionBody funcBody = (BLangBlockFunctionBody) packageNode.initFunction.body;
-        BLangFunction initFunction = packageNode.initFunction;
-
-        List<BLangFunction> generatedFunctions = new ArrayList<>();
-        List<BLangStatement> stmts = new ArrayList<>(funcBody.stmts);
-        funcBody.stmts.clear();
-        BLangFunction newFunc = initFunction;
-        BLangBlockFunctionBody newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-
-        // until we get to a varDef, stmts are independent, divide it based on methodSize
-        int varDefIndex = 0;
-        for (int i = 0; i < stmts.size(); i++) {
-            BLangStatement statement = stmts.get(i);
-            if (statement.getKind() == NodeKind.VARIABLE_DEF) {
-                break;
-            }
-            varDefIndex++;
-            if (i > 0 && (i % methodSize == 0 || isAssignmentWithInitOrRecordLiteralExpr(statement))) {
-                generatedFunctions.add(newFunc);
-                newFunc = createIntermediateInitFunction(packageNode, env);
-                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
-            }
-
-            newFuncBody.stmts.add(stmts.get(i));
-        }
-
-        // from a varDef to a service constructor, those stmts should be within single method
-        List<BLangStatement> chunkStmts = new ArrayList<>();
-        for (int i = varDefIndex; i < stmts.size(); i++) {
-            BLangStatement stmt = stmts.get(i);
-            chunkStmts.add(stmt);
-            varDefIndex++;
-            if ((stmt.getKind() == NodeKind.ASSIGNMENT) &&
-                    (((BLangAssignment) stmt).expr.getKind() == NodeKind.SERVICE_CONSTRUCTOR) &&
-                    (newFuncBody.stmts.size() + chunkStmts.size() > methodSize)) {
-                // enf of current chunk
-                if (newFuncBody.stmts.size() + chunkStmts.size() > methodSize) {
-                    generatedFunctions.add(newFunc);
-                    newFunc = createIntermediateInitFunction(packageNode, env);
-                    newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-                    symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
-                }
-                newFuncBody.stmts.addAll(chunkStmts);
-                chunkStmts.clear();
-            } else if ((stmt.getKind() == NodeKind.ASSIGNMENT) &&
-                    (((BLangAssignment) stmt).varRef instanceof BLangPackageVarRef) &&
-                    Symbols.isFlagOn(((BLangPackageVarRef) ((BLangAssignment) stmt).varRef).varSymbol.flags,
-                            Flags.LISTENER)
-            ) {
-                // this is where listener registrations starts, they are independent stmts
-                break;
-            }
-        }
-        newFuncBody.stmts.addAll(chunkStmts);
-
-        // rest of the statements can be split without chunks
-        for (int i = varDefIndex; i < stmts.size(); i++) {
-            if (i > 0 && i % methodSize == 0) {
-                generatedFunctions.add(newFunc);
-                newFunc = createIntermediateInitFunction(packageNode, env);
-                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
-            }
-            newFuncBody.stmts.add(stmts.get(i));
-        }
-
-        generatedFunctions.add(newFunc);
-
-        for (int j = 0; j < generatedFunctions.size() - 1; j++) {
-            BLangFunction thisFunction = generatedFunctions.get(j);
-
-            BLangCheckedExpr checkedExpr =
-                    ASTBuilderUtil.createCheckExpr(initFunction.pos,
-                                                   createInvocationNode(generatedFunctions.get(j + 1).name.value,
-                                                                        new ArrayList<>(), symTable.errorOrNilType),
-                                                   symTable.nilType);
-            checkedExpr.equivalentErrorTypeList.add(symTable.errorType);
-
-            BLangExpressionStmt expressionStmt = ASTBuilderUtil
-                    .createExpressionStmt(thisFunction.pos, (BLangBlockFunctionBody) thisFunction.body);
-            expressionStmt.expr = checkedExpr;
-            expressionStmt.expr.pos = initFunction.pos;
-
-            if (j > 0) { // skip init func
-                thisFunction = rewrite(thisFunction, env);
-                packageNode.functions.add(thisFunction);
-                packageNode.topLevelNodes.add(thisFunction);
-            }
-        }
-
-        rewriteLastSplitFunction(packageNode, env, generatedFunctions);
-        return generatedFunctions.get(0);
-    }
-
-    private BLangFunction splitStartFunction(BLangPackage packageNode, SymbolEnv env) {
-        BLangBlockFunctionBody funcBody = (BLangBlockFunctionBody) packageNode.startFunction.body;
-        BLangFunction startFunction = packageNode.startFunction;
-
-        List<BLangFunction> generatedFunctions = new ArrayList<>();
-        List<BLangStatement> stmts = new ArrayList<>(funcBody.stmts);
-        funcBody.stmts.clear();
-        BLangFunction newFunc = startFunction;
-        BLangBlockFunctionBody newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-
-        for (int i = 0; i < stmts.size() - 1; i++) {
-            if (i > 0 && (i % LISTENER_COUNT_PER_METHOD == 0)) {
-                generatedFunctions.add(newFunc);
-                newFunc = createIntermediateStartFunction(packageNode, env);
-                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
-            }
-            newFuncBody.stmts.add(stmts.get(i));
-        }
-
-        newFuncBody.stmts.add(stmts.get(stmts.size() - 1));
-        generatedFunctions.add(newFunc);
-
-        for (int j = 0; j < generatedFunctions.size() - 1; j++) {
-            BLangFunction thisFunction = generatedFunctions.get(j);
-
-            BLangCheckedExpr checkedExpr =
-                    ASTBuilderUtil.createCheckExpr(startFunction.pos,
-                            createInvocationNode(generatedFunctions.get(j + 1).name.value,
-                                    new ArrayList<>(), symTable.errorOrNilType),
-                            symTable.nilType);
-            checkedExpr.equivalentErrorTypeList.add(symTable.errorType);
-
-            BLangExpressionStmt expressionStmt = ASTBuilderUtil
-                    .createExpressionStmt(thisFunction.pos, (BLangBlockFunctionBody) thisFunction.body);
-            expressionStmt.expr = checkedExpr;
-            expressionStmt.expr.pos = startFunction.pos;
-
-            if (j > 0) { // skip start func
-                thisFunction = rewrite(thisFunction, env);
-                packageNode.functions.add(thisFunction);
-                packageNode.topLevelNodes.add(thisFunction);
-            }
-        }
-
-        rewriteLastSplitFunction(packageNode, env, generatedFunctions);
-        return generatedFunctions.get(0);
-    }
-
-    private void rewriteLastSplitFunction(BLangPackage packageNode, SymbolEnv env,
-                                          List<BLangFunction> generatedFunctions) {
-        if (generatedFunctions.size() > 1) {
-            // add last func
-            BLangFunction lastFunc = generatedFunctions.get(generatedFunctions.size() - 1);
-            lastFunc = rewrite(lastFunc, env);
-            packageNode.functions.add(lastFunc);
-            packageNode.topLevelNodes.add(lastFunc);
-        }
-    }
-
-    private BLangFunction splitStopFunction(BLangPackage packageNode, SymbolEnv env) {
-        BLangBlockFunctionBody funcBody = (BLangBlockFunctionBody) packageNode.stopFunction.body;
-        BLangFunction stopFunction = packageNode.stopFunction;
-
-        List<BLangFunction> generatedFunctions = new ArrayList<>();
-        List<BLangStatement> stmts = new ArrayList<>(funcBody.stmts);
-        funcBody.stmts.clear();
-        BLangFunction newFunc = stopFunction;
-        BLangBlockFunctionBody newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-
-        for (int i = 0; i < stmts.size() - 1; i++) {
-            if (i > 0 && (i % LISTENER_COUNT_PER_METHOD == 0)) {
-                generatedFunctions.add(newFunc);
-                newFunc = createIntermediateStopFunction(packageNode, env);
-                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
-                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
-            }
-            newFuncBody.stmts.add(stmts.get(i));
-        }
-
-        newFuncBody.stmts.add(stmts.get(stmts.size() - 1));
-        generatedFunctions.add(newFunc);
-
-        for (int j = 0; j < generatedFunctions.size() - 1; j++) {
-            BLangFunction thisFunction = generatedFunctions.get(j);
-
-            BInvokableSymbol invokableSymbol = (BInvokableSymbol) symTable.rootScope.lookup(
-                    new Name(generatedFunctions.get(j + 1).name.value)).symbol;
-            BLangInvocation invocationExpr = ASTBuilderUtil.createInvocationExpr(stopFunction.pos, invokableSymbol,
-                    new ArrayList<>(), symResolver);
-
-            BLangExpressionStmt expressionStmt = ASTBuilderUtil
-                    .createExpressionStmt(thisFunction.pos, (BLangBlockFunctionBody) thisFunction.body);
-            expressionStmt.expr = invocationExpr;
-            expressionStmt.expr.pos = stopFunction.pos;
-
-            if (j > 0) { // skip start func
-                thisFunction = rewrite(thisFunction, env);
-                packageNode.functions.add(thisFunction);
-                packageNode.topLevelNodes.add(thisFunction);
-            }
-        }
-
-        rewriteLastSplitFunction(packageNode, env, generatedFunctions);
-        return generatedFunctions.get(0);
-    }
-
-    private boolean isAssignmentWithInitOrRecordLiteralExpr(BLangStatement statement) {
-        if (statement.getKind() == NodeKind.ASSIGNMENT) {
-            return isMappingOrObjectConstructorOrObjInit(((BLangAssignment) statement).getExpression());
-        }
-        return false;
-    }
-
     protected boolean isMappingOrObjectConstructorOrObjInit(BLangExpression expression) {
         switch (expression.getKind()) {
             case TYPE_INIT_EXPR:
@@ -10250,45 +10026,6 @@ public class Desugar extends BLangNodeVisitor {
             default:
                 return false;
         }
-    }
-
-    /**
-     * Create an intermediate package init function.
-     *
-     * @param pkgNode package node
-     * @param env     symbol environment of package
-     */
-    private BLangFunction createIntermediateInitFunction(BLangPackage pkgNode, SymbolEnv env) {
-        String alias = pkgNode.symbol.pkgID.toString();
-        BLangFunction initFunction = ASTBuilderUtil
-                .createInitFunctionWithErrorOrNilReturn(pkgNode.pos, alias,
-                                                        new Name(Names.INIT_FUNCTION_SUFFIX.value
-                                                                + this.initFuncIndex++), symTable);
-        // Create invokable symbol for init function
-        createInvokableSymbol(initFunction, env);
-        return initFunction;
-    }
-
-    private BLangFunction createIntermediateStartFunction(BLangPackage pkgNode, SymbolEnv env) {
-        String alias = pkgNode.symbol.pkgID.toString();
-        BLangFunction startFunction = ASTBuilderUtil
-                .createInitFunctionWithErrorOrNilReturn(new BLangDiagnosticLocation("$_start", 0, 0, 0, 0), alias,
-                        new Name(Names.START_FUNCTION_SUFFIX.value
-                                + this.startFuncIndex++), symTable);
-        // Create invokable symbol for start function
-        createInvokableSymbol(startFunction, env);
-        return startFunction;
-    }
-
-    private BLangFunction createIntermediateStopFunction(BLangPackage pkgNode, SymbolEnv env) {
-        String alias = pkgNode.symbol.pkgID.toString();
-        BLangFunction stopFunction = ASTBuilderUtil
-                .createInitFunctionWithNilReturn(new BLangDiagnosticLocation("$_stop", 0, 0, 0, 0), alias,
-                        new Name(Names.STOP_FUNCTION_SUFFIX.value
-                                + this.stopFuncIndex++));
-        // Create invokable symbol for stop function
-        createInvokableSymbol(stopFunction, env);
-        return stopFunction;
     }
 
     private BType getRestType(BInvokableSymbol invokableSymbol) {
