+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/desugar/LargeMethodSplitter.java
@@ -1,0 +1,379 @@
+/*
+ *  Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  Version 2.0 (the "License"); you may not use this file except
+ *  in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing,
+ *  software distributed under the License is distributed on an
+ *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  KIND, either express or implied.  See the License for the
+ *  specific language governing permissions and limitations
+ *  under the License.
+ */
+
+package org.wso2.ballerinalang.compiler.desugar;
+
+import org.ballerinalang.model.tree.NodeKind;
+import org.wso2.ballerinalang.compiler.diagnostic.BLangDiagnosticLocation;
+import org.wso2.ballerinalang.compiler.semantics.analyzer.SymbolResolver;
+import org.wso2.ballerinalang.compiler.semantics.model.SymbolEnv;
+import org.wso2.ballerinalang.compiler.semantics.model.SymbolTable;
+import org.wso2.ballerinalang.compiler.semantics.model.symbols.BInvokableSymbol;
+import org.wso2.ballerinalang.compiler.semantics.model.symbols.Symbols;
+import org.wso2.ballerinalang.compiler.tree.BLangBlockFunctionBody;
+import org.wso2.ballerinalang.compiler.tree.BLangFunction;
+import org.wso2.ballerinalang.compiler.tree.BLangPackage;
+import org.wso2.ballerinalang.compiler.tree.expressions.BLangCheckedExpr;
+import org.wso2.ballerinalang.compiler.tree.expressions.BLangInvocation;
+import org.wso2.ballerinalang.compiler.tree.expressions.BLangSimpleVarRef;
+import org.wso2.ballerinalang.compiler.tree.statements.BLangAssignment;
+import org.wso2.ballerinalang.compiler.tree.statements.BLangExpressionStmt;
+import org.wso2.ballerinalang.compiler.tree.statements.BLangStatement;
+import org.wso2.ballerinalang.compiler.util.CompilerContext;
+import org.wso2.ballerinalang.compiler.util.Name;
+import org.wso2.ballerinalang.compiler.util.Names;
+import org.wso2.ballerinalang.util.Flags;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.wso2.ballerinalang.compiler.util.Constants.INIT_METHOD_SPLIT_SIZE;
+import static org.wso2.ballerinalang.compiler.util.Constants.LISTENER_COUNT_PER_METHOD;
+
+/**
+ * Split large init, start, stop BLangFunctions into several smaller functions.
+ *
+ * @since 2201.1.0
+ */
+public class LargeMethodSplitter {
+
+    private static final String SPLIT_START_FUNCTIONS_CLASS_NAME = "$_start";
+    private static final String SPLIT_STOP_FUNCTIONS_CLASS_NAME = "$_stop";
+    private int initFuncIndex = 0;
+    private int startFuncIndex = 0;
+    private int stopFuncIndex = 0;
+
+    private final Desugar desugar;
+    private final SymbolTable symTable;
+    private final Names names;
+    private final SymbolResolver symResolver;
+
+    private static final CompilerContext.Key<LargeMethodSplitter> LARGE_METHOD_SPLITTER_KEY =
+            new CompilerContext.Key<>();
+
+    public static LargeMethodSplitter getInstance(CompilerContext context) {
+        LargeMethodSplitter largeMethodSplitter = context.get(LARGE_METHOD_SPLITTER_KEY);
+        if (largeMethodSplitter == null) {
+            largeMethodSplitter = new LargeMethodSplitter(context);
+        }
+        return largeMethodSplitter;
+    }
+
+    private LargeMethodSplitter(CompilerContext context) {
+        context.put(LARGE_METHOD_SPLITTER_KEY, this);
+        this.desugar = Desugar.getInstance(context);
+        this.symTable = SymbolTable.getInstance(context);
+        this.names = Names.getInstance(context);
+        this.symResolver = SymbolResolver.getInstance(context);
+    }
+
+    void splitLargeFunctions(BLangPackage pkgNode, SymbolEnv env) {
+        pkgNode.initFunction = splitInitFunction(pkgNode, env);
+        pkgNode.startFunction = splitStartFunction(pkgNode, env);
+        pkgNode.stopFunction = splitStopFunction(pkgNode, env);
+    }
+
+    /**
+     * Split package init function into several smaller functions.
+     *
+     * @param packageNode package node
+     * @param env symbol environment
+     * @return initial init function but trimmed in size
+     */
+    private BLangFunction splitInitFunction(BLangPackage packageNode, SymbolEnv env) {
+        int methodSize = INIT_METHOD_SPLIT_SIZE;
+        BLangBlockFunctionBody funcBody = (BLangBlockFunctionBody) packageNode.initFunction.body;
+        BLangFunction initFunction = packageNode.initFunction;
+
+        List<BLangFunction> generatedFunctions = new ArrayList<>();
+        List<BLangStatement> stmts = new ArrayList<>(funcBody.stmts);
+        funcBody.stmts.clear();
+        BLangFunction newFunc = initFunction;
+        BLangBlockFunctionBody newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+
+        // until we get to a varDef, stmts are independent, divide it based on methodSize
+        int varDefIndex = 0;
+        for (int i = 0; i < stmts.size(); i++) {
+            BLangStatement statement = stmts.get(i);
+            if (statement.getKind() == NodeKind.VARIABLE_DEF) {
+                break;
+            }
+            varDefIndex++;
+            if (i > 0 && (i % methodSize == 0 || isAssignmentWithInitOrRecordLiteralExpr(statement))) {
+                generatedFunctions.add(newFunc);
+                newFunc = createIntermediateInitFunction(packageNode, env);
+                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
+            }
+            newFuncBody.stmts.add(stmts.get(i));
+        }
+
+        // from a varDef to a service constructor, those stmts should be within single method
+        List<BLangStatement> chunkStmts = new ArrayList<>();
+        for (int i = varDefIndex; i < stmts.size(); i++) {
+            BLangStatement stmt = stmts.get(i);
+            chunkStmts.add(stmt);
+            varDefIndex++;
+            if ((stmt.getKind() == NodeKind.ASSIGNMENT) &&
+                    (((BLangAssignment) stmt).expr.getKind() == NodeKind.SERVICE_CONSTRUCTOR) &&
+                    (newFuncBody.stmts.size() + chunkStmts.size() > methodSize)) {
+                // enf of current chunk
+                if (newFuncBody.stmts.size() + chunkStmts.size() > methodSize) {
+                    generatedFunctions.add(newFunc);
+                    newFunc = createIntermediateInitFunction(packageNode, env);
+                    newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+                    symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
+                }
+                newFuncBody.stmts.addAll(chunkStmts);
+                chunkStmts.clear();
+            } else if ((stmt.getKind() == NodeKind.ASSIGNMENT) &&
+                    (((BLangAssignment) stmt).varRef instanceof BLangSimpleVarRef.BLangPackageVarRef) &&
+                    Symbols.isFlagOn(
+                            ((BLangSimpleVarRef.BLangPackageVarRef) ((BLangAssignment) stmt).varRef).varSymbol.flags,
+                            Flags.LISTENER)
+            ) {
+                // this is where listener registrations starts, they are independent stmts
+                break;
+            }
+        }
+        newFuncBody.stmts.addAll(chunkStmts);
+
+        // rest of the statements can be split without chunks
+        for (int i = varDefIndex; i < stmts.size(); i++) {
+            if (i > 0 && i % methodSize == 0) {
+                generatedFunctions.add(newFunc);
+                newFunc = createIntermediateInitFunction(packageNode, env);
+                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
+            }
+            newFuncBody.stmts.add(stmts.get(i));
+        }
+
+        generatedFunctions.add(newFunc);
+
+        for (int j = 0; j < generatedFunctions.size() - 1; j++) {
+            BLangFunction thisFunction = generatedFunctions.get(j);
+
+            BLangCheckedExpr checkedExpr =
+                    ASTBuilderUtil.createCheckExpr(initFunction.pos,
+                            desugar.createInvocationNode(generatedFunctions.get(j + 1).name.value,
+                                    new ArrayList<>(), symTable.errorOrNilType),
+                            symTable.nilType);
+            checkedExpr.equivalentErrorTypeList.add(symTable.errorType);
+
+            BLangExpressionStmt expressionStmt = ASTBuilderUtil
+                    .createExpressionStmt(thisFunction.pos, (BLangBlockFunctionBody) thisFunction.body);
+            expressionStmt.expr = checkedExpr;
+            expressionStmt.expr.pos = initFunction.pos;
+
+            if (j > 0) { // skip init func
+                thisFunction = desugar.rewrite(thisFunction, env);
+                packageNode.functions.add(thisFunction);
+                packageNode.topLevelNodes.add(thisFunction);
+            }
+        }
+
+        rewriteLastSplitFunction(packageNode, env, generatedFunctions);
+        initFuncIndex = 0;
+        return generatedFunctions.get(0);
+    }
+
+    private boolean isAssignmentWithInitOrRecordLiteralExpr(BLangStatement statement) {
+        if (statement.getKind() == NodeKind.ASSIGNMENT) {
+            return desugar.isMappingOrObjectConstructorOrObjInit(((BLangAssignment) statement).getExpression());
+        }
+        return false;
+    }
+
+    /**
+     * Split package start function into several smaller functions.
+     *
+     * @param packageNode package node
+     * @param env symbol environment
+     * @return initial start function but trimmed in size
+     */
+    private BLangFunction splitStartFunction(BLangPackage packageNode, SymbolEnv env) {
+        BLangBlockFunctionBody funcBody = (BLangBlockFunctionBody) packageNode.startFunction.body;
+        BLangFunction startFunction = packageNode.startFunction;
+
+        List<BLangFunction> generatedFunctions = new ArrayList<>();
+        List<BLangStatement> stmts = new ArrayList<>(funcBody.stmts);
+        funcBody.stmts.clear();
+        BLangFunction newFunc = startFunction;
+        BLangBlockFunctionBody newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+
+        for (int i = 0; i < stmts.size() - 1; i++) {
+            if (i > 0 && (i % LISTENER_COUNT_PER_METHOD == 0)) {
+                generatedFunctions.add(newFunc);
+                newFunc = createIntermediateStartFunction(packageNode, env);
+                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
+            }
+            newFuncBody.stmts.add(stmts.get(i));
+        }
+
+        newFuncBody.stmts.add(stmts.get(stmts.size() - 1));
+        generatedFunctions.add(newFunc);
+
+        for (int j = 0; j < generatedFunctions.size() - 1; j++) {
+            BLangFunction thisFunction = generatedFunctions.get(j);
+
+            BLangCheckedExpr checkedExpr =
+                    ASTBuilderUtil.createCheckExpr(startFunction.pos,
+                            desugar.createInvocationNode(generatedFunctions.get(j + 1).name.value,
+                                    new ArrayList<>(), symTable.errorOrNilType),
+                            symTable.nilType);
+            checkedExpr.equivalentErrorTypeList.add(symTable.errorType);
+
+            BLangExpressionStmt expressionStmt = ASTBuilderUtil
+                    .createExpressionStmt(thisFunction.pos, (BLangBlockFunctionBody) thisFunction.body);
+            expressionStmt.expr = checkedExpr;
+            expressionStmt.expr.pos = startFunction.pos;
+
+            if (j > 0) { // skip start func
+                thisFunction = desugar.rewrite(thisFunction, env);
+                packageNode.functions.add(thisFunction);
+                packageNode.topLevelNodes.add(thisFunction);
+            }
+        }
+
+        rewriteLastSplitFunction(packageNode, env, generatedFunctions);
+        startFuncIndex = 0;
+        return generatedFunctions.get(0);
+    }
+
+    /**
+     * Split package stop function into several smaller functions.
+     *
+     * @param packageNode package node
+     * @param env symbol environment
+     * @return initial stop function but trimmed in size
+     */
+    private BLangFunction splitStopFunction(BLangPackage packageNode, SymbolEnv env) {
+        BLangBlockFunctionBody funcBody = (BLangBlockFunctionBody) packageNode.stopFunction.body;
+        BLangFunction stopFunction = packageNode.stopFunction;
+
+        List<BLangFunction> generatedFunctions = new ArrayList<>();
+        List<BLangStatement> stmts = new ArrayList<>(funcBody.stmts);
+        funcBody.stmts.clear();
+        BLangFunction newFunc = stopFunction;
+        BLangBlockFunctionBody newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+
+        for (int i = 0; i < stmts.size() - 1; i++) {
+            if (i > 0 && (i % LISTENER_COUNT_PER_METHOD == 0)) {
+                generatedFunctions.add(newFunc);
+                newFunc = createIntermediateStopFunction(packageNode, env);
+                newFuncBody = (BLangBlockFunctionBody) newFunc.body;
+                symTable.rootScope.define(names.fromIdNode(newFunc.name), newFunc.symbol);
+            }
+            newFuncBody.stmts.add(stmts.get(i));
+        }
+
+        newFuncBody.stmts.add(stmts.get(stmts.size() - 1));
+        generatedFunctions.add(newFunc);
+
+        for (int j = 0; j < generatedFunctions.size() - 1; j++) {
+            BLangFunction thisFunction = generatedFunctions.get(j);
+
+            BInvokableSymbol invokableSymbol = (BInvokableSymbol) symTable.rootScope.lookup(
+                    new Name(generatedFunctions.get(j + 1).name.value)).symbol;
+            BLangInvocation invocationExpr = ASTBuilderUtil.createInvocationExpr(stopFunction.pos, invokableSymbol,
+                    new ArrayList<>(), symResolver);
+
+            BLangExpressionStmt expressionStmt = ASTBuilderUtil
+                    .createExpressionStmt(thisFunction.pos, (BLangBlockFunctionBody) thisFunction.body);
+            expressionStmt.expr = invocationExpr;
+            expressionStmt.expr.pos = stopFunction.pos;
+
+            if (j > 0) { // skip start func
+                thisFunction = desugar.rewrite(thisFunction, env);
+                packageNode.functions.add(thisFunction);
+                packageNode.topLevelNodes.add(thisFunction);
+            }
+        }
+
+        rewriteLastSplitFunction(packageNode, env, generatedFunctions);
+        stopFuncIndex = 0;
+        return generatedFunctions.get(0);
+    }
+
+    private void rewriteLastSplitFunction(BLangPackage packageNode, SymbolEnv env,
+                                          List<BLangFunction> generatedFunctions) {
+        if (generatedFunctions.size() > 1) {
+            // add last func
+            BLangFunction lastFunc = generatedFunctions.get(generatedFunctions.size() - 1);
+            lastFunc = desugar.rewrite(lastFunc, env);
+            packageNode.functions.add(lastFunc);
+            packageNode.topLevelNodes.add(lastFunc);
+        }
+    }
+
+    /**
+     * Create an intermediate package init function.
+     *
+     * @param pkgNode package node
+     * @param env     symbol environment of package
+     * @return created intermediate init function
+     */
+    private BLangFunction createIntermediateInitFunction(BLangPackage pkgNode, SymbolEnv env) {
+        String alias = pkgNode.symbol.pkgID.toString();
+        BLangFunction initFunction = ASTBuilderUtil
+                .createInitFunctionWithErrorOrNilReturn(pkgNode.pos, alias,
+                        new Name(Names.INIT_FUNCTION_SUFFIX.value
+                                + this.initFuncIndex++), symTable);
+        // Create invokable symbol for init function
+        desugar.createInvokableSymbol(initFunction, env);
+        return initFunction;
+    }
+
+    /**
+     * Create an intermediate package start function.
+     *
+     * @param pkgNode package node
+     * @param env     symbol environment of package
+     * @return created intermediate start function
+     */
+    private BLangFunction createIntermediateStartFunction(BLangPackage pkgNode, SymbolEnv env) {
+        String alias = pkgNode.symbol.pkgID.toString();
+        BLangFunction startFunction = ASTBuilderUtil
+                .createInitFunctionWithErrorOrNilReturn(new BLangDiagnosticLocation(SPLIT_START_FUNCTIONS_CLASS_NAME,
+                                0, 0, 0, 0, 0, 0),
+                        alias, new Name(Names.START_FUNCTION_SUFFIX.value + this.startFuncIndex++), symTable);
+        // Create invokable symbol for start function
+        desugar.createInvokableSymbol(startFunction, env);
+        return startFunction;
+    }
+
+    /**
+     * Create an intermediate package stop function.
+     *
+     * @param pkgNode package node
+     * @param env     symbol environment of package
+     * @return created intermediate stop function
+     */
+    private BLangFunction createIntermediateStopFunction(BLangPackage pkgNode, SymbolEnv env) {
+        String alias = pkgNode.symbol.pkgID.toString();
+        BLangFunction stopFunction = ASTBuilderUtil
+                .createInitFunctionWithNilReturn(new BLangDiagnosticLocation(SPLIT_STOP_FUNCTIONS_CLASS_NAME,
+                                0, 0, 0, 0, 0, 0),
+                        alias, new Name(Names.STOP_FUNCTION_SUFFIX.value + this.stopFuncIndex++));
+        // Create invokable symbol for stop function
+        desugar.createInvokableSymbol(stopFunction, env);
+        return stopFunction;
+    }
+}
